Added docker package
